wolfactiv_mbti_api/requirements.txt
openai
python-dotenv
supabase
fastapi
uvicorn
